% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_fuzzy_join_cpp.R
\name{fstring_inner_join}
\alias{fstring_inner_join}
\title{Fuzzy inner join}
\usage{
fstring_inner_join(x, y, by = NULL, match_fun, ...)
}
\arguments{
\item{x}{A \code{data.frame} or \code{data.table}.}

\item{y}{A \code{data.frame} or \code{data.table}.}

\item{by}{Columns by which to join the two tables. See
\code{\link{fuzzystring_join}}.}

\item{match_fun}{A function used to match values. It must return a logical
vector (or a data.frame/data.table whose first column is logical) indicating
which pairs match. For multi-column joins, you may pass a list of functions
(one per column).}

\item{...}{Additional arguments passed to the matching function(s).}
}
\value{
See \code{\link{fuzzystring_join_backend}}.
}
\description{
Convenience wrapper for \code{fuzzystring_join_backend(mode = "inner")}.
}
