% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzydt-package.R
\docType{package}
\name{fuzzystring-package}
\alias{fuzzystring}
\alias{fuzzystring-package}
\title{fuzzystring: Fast Fuzzy String Joins for Data Frames}
\description{
Perform fuzzy joins on data frames using approximate string matching. Implements all standard join types (inner, left, right, full, semi, anti) with support for multiple string distance metrics from the 'stringdist' package including Levenshtein, Damerau-Levenshtein, Jaro-Winkler, and Soundex. Features a high-performance 'data.table' backend with 'C++' row binding for efficient processing of large datasets. Ideal for matching misspellings, inconsistent labels, messy user input, or reconciling datasets with slight variations in identifiers. Optionally returns distance metrics alongside matched records.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PaulESantos/fuzzystring}
  \item \url{https://paulesantos.github.io/fuzzystring/}
  \item Report bugs at \url{https://github.com/PaulESantos/fuzzystring/issues}
}

}
\author{
\strong{Maintainer}: Paul E. Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID})

Other contributors:
\itemize{
  \item David Robinson \email{admiral.david@gmail.com} (aut of fuzzyjoin) [contributor]
}

}
\keyword{internal}
