% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzystring-package.R
\docType{package}
\name{fuzzystring}
\alias{fuzzystring-package}
\alias{fuzzystring}
\title{fuzzystring: Fast fuzzy string joins for data frames}
\description{
\code{fuzzystring} provides fuzzy inner, left, right, full, semi, and anti joins
for \code{data.frame} and \code{data.table} objects using approximate string matching.
It combines \code{stringdist} metrics with a \code{data.table} backend and compiled 'C++'
result assembly to reduce overhead in large joins while preserving familiar
join semantics.
}
\details{
Main entry points are \code{\link[=fuzzystring_join]{fuzzystring_join()}} and the convenience wrappers
\code{\link[=fuzzystring_inner_join]{fuzzystring_inner_join()}}, \code{\link[=fuzzystring_left_join]{fuzzystring_left_join()}},
\code{\link[=fuzzystring_right_join]{fuzzystring_right_join()}}, \code{\link[=fuzzystring_full_join]{fuzzystring_full_join()}},
\code{\link[=fuzzystring_semi_join]{fuzzystring_semi_join()}}, and \code{\link[=fuzzystring_anti_join]{fuzzystring_anti_join()}}.

The package also includes the example dataset \link{misspellings}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PaulESantos/fuzzystring}
  \item \url{https://paulesantos.github.io/fuzzystring/}
  \item Report bugs at \url{https://github.com/PaulESantos/fuzzystring/issues}
}

}
\author{
\strong{Maintainer}: Paul E. Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Paul E. Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item David Robinson \email{admiral.david@gmail.com} (aut of fuzzyjoin) [contributor]
}

}
\keyword{package}
