% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzydt-package.R
\docType{package}
\name{fuzzystring-package}
\alias{fuzzystring}
\alias{fuzzystring-package}
\title{fuzzystring: Fuzzy String Joins for Data Frames}
\description{
Join data frames using approximate string matching. Provides inner/left/right/full/semi/anti joins powered by string distance metrics from 'stringdist', with a fast data.table backend and optional distance columns in the output.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PaulESantos/fuzzystring}
  \item \url{https://paulesantos.github.io/fuzzystring/}
  \item Report bugs at \url{https://github.com/PaulESantos/fuzzystring/issues}
}

}
\author{
\strong{Maintainer}: Paul E. Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID})

}
\keyword{internal}
